+++ crates/blockifier/src/execution.rs
@@ -1,3 +1,4 @@
+pub mod bouncer;
 pub mod call_info;
 pub mod common_hints;
 pub mod contract_address;
